daemon: move snapshot cmd and tests into own files
Also makes snapshot tests be in daemon_test package, starting down the
road of updating daemon to our current practices.

Also², make tests better by following @mvo5's suggestions (thanks!).
