+++ daemon/api.go
@@ -259,15 +259,6 @@ var (
 		POST:     ackWarnings,
 	}
 
-	snapshotCmd = &Command{
-		// TODO: also support /v2/snapshots/<id>
-		Path:     "/v2/snapshots",
-		UserOK:   true,
-		PolkitOK: "io.snapcraft.snapd.manage",
-		GET:      listSnapshots,
-		POST:     changeSnapshots,
-	}
-
 	buildID = "unknown"
 )
 
@@ -919,10 +910,10 @@ func (inst *snapInstruction) installFlags() (snapstate.Flags, error) {
 }
 
 type snapInstructionResult struct {
-	summary  string
-	affected []string
-	tasksets []*state.TaskSet
-	result   map[string]interface{}
+	Summary  string
+	Affected []string
+	Tasksets []*state.TaskSet
+	Result   map[string]interface{}
 }
 
 var (
@@ -1006,9 +997,9 @@ func snapUpdateMany(inst *snapInstruction, st *state.State) (*snapInstructionRes
 	}
 
 	return &snapInstructionResult{
-		summary:  msg,
-		affected: updated,
-		tasksets: tasksets,
+		Summary:  msg,
+		Affected: updated,
+		Tasksets: tasksets,
 	}, nil
 }
 
@@ -1051,9 +1042,9 @@ func snapInstallMany(inst *snapInstruction, st *state.State) (*snapInstructionRe
 	}
 
 	return &snapInstructionResult{
-		summary:  msg,
-		affected: installed,
-		tasksets: tasksets,
+		Summary:  msg,
+		Affected: installed,
+		Tasksets: tasksets,
 	}, nil
 }
 
@@ -1131,9 +1122,9 @@ func snapRemoveMany(inst *snapInstruction, st *state.State) (*snapInstructionRes
 	}
 
 	return &snapInstructionResult{
-		summary:  msg,
-		affected: removed,
-		tasksets: tasksets,
+		Summary:  msg,
+		Affected: removed,
+		Tasksets: tasksets,
 	}, nil
 }
 
@@ -1222,10 +1213,10 @@ func snapshotMany(inst *snapInstruction, st *state.State) (*snapInstructionResul
 	}
 
 	return &snapInstructionResult{
-		summary:  msg,
-		affected: snapshotted,
-		tasksets: []*state.TaskSet{ts},
-		result:   map[string]interface{}{"set-id": setID},
+		Summary:  msg,
+		Affected: snapshotted,
+		Tasksets: []*state.TaskSet{ts},
+		Result:   map[string]interface{}{"set-id": setID},
 	}, nil
 }
 
@@ -1412,17 +1403,17 @@ func snapsOp(c *Command, r *http.Request, user *auth.UserState) Response {
 	}
 
 	var chg *state.Change
-	if len(res.tasksets) == 0 {
-		chg = st.NewChange(inst.Action+"-snap", res.summary)
+	if len(res.Tasksets) == 0 {
+		chg = st.NewChange(inst.Action+"-snap", res.Summary)
 		chg.SetStatus(state.DoneStatus)
 	} else {
-		chg = newChange(st, inst.Action+"-snap", res.summary, res.tasksets, res.affected)
+		chg = newChange(st, inst.Action+"-snap", res.Summary, res.Tasksets, res.Affected)
 		ensureStateSoon(st)
 	}
 
-	chg.Set("api-data", map[string]interface{}{"snap-names": res.affected})
+	chg.Set("api-data", map[string]interface{}{"snap-names": res.Affected})
 
-	return AsyncResponse(res.result, &Meta{Change: chg.ID()})
+	return AsyncResponse(res.Result, &Meta{Change: chg.ID()})
 }
 
 func postSnaps(c *Command, r *http.Request, user *auth.UserState) Response {
@@ -2960,100 +2951,4 @@ func getWarnings(c *Command, r *http.Request, _ *auth.UserState) Response {
 	}
 
 	return SyncResponse(ws, nil)
-}
-
-func listSnapshots(c *Command, r *http.Request, user *auth.UserState) Response {
-	query := r.URL.Query()
-	var setID uint64
-	if sid := query.Get("set"); sid != "" {
-		var err error
-		setID, err = strconv.ParseUint(sid, 10, 64)
-		if err != nil {
-			return BadRequest("'set', if given, must be a positive base 10 number; got %q", sid)
-		}
-	}
-
-	sets, err := snapshotList(context.TODO(), setID, splitQS(r.URL.Query().Get("snaps")))
-	if err != nil {
-		return InternalError("%v", err)
-	}
-	return SyncResponse(sets, nil)
-}
-
-// A snapshotAction is used to request an operation on a snapshot
-// keep this in sync with client/snapshotAction...
-type snapshotAction struct {
-	SetID  uint64   `json:"set"`
-	Action string   `json:"action"`
-	Snaps  []string `json:"snaps,omitempty"`
-	Users  []string `json:"users,omitempty"`
-}
-
-func (action snapshotAction) String() string {
-	// verb of snapshot #N [for snaps %q] [for users %q]
-	var snaps string
-	var users string
-	if len(action.Snaps) > 0 {
-		snaps = " for snaps " + strutil.Quoted(action.Snaps)
-	}
-	if len(action.Users) > 0 {
-		users = " for users " + strutil.Quoted(action.Users)
-	}
-	return fmt.Sprintf("%s of snapshot set #%d%s%s", strings.Title(action.Action), action.SetID, snaps, users)
-}
-
-func changeSnapshots(c *Command, r *http.Request, user *auth.UserState) Response {
-	var action snapshotAction
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&action); err != nil {
-		return BadRequest("cannot decode request body into snapshot operation: %v", err)
-	}
-	if decoder.More() {
-		return BadRequest("extra content found after snapshot operation")
-	}
-
-	if action.SetID == 0 {
-		return BadRequest("snapshot operation requires snapshot set ID")
-	}
-
-	if action.Action == "" {
-		return BadRequest("snapshot operation requires action")
-	}
-
-	var affected []string
-	var ts *state.TaskSet
-	var err error
-
-	st := c.d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-
-	switch action.Action {
-	case "check":
-		affected, ts, err = snapshotCheck(st, action.SetID, action.Snaps, action.Users)
-	case "restore":
-		affected, ts, err = snapshotRestore(st, action.SetID, action.Snaps, action.Users)
-	case "forget":
-		if len(action.Users) != 0 {
-			return BadRequest(`snapshot "forget" operation cannot specify users`)
-		}
-		affected, ts, err = snapshotForget(st, action.SetID, action.Snaps)
-	default:
-		return BadRequest("unknown snapshot operation %q", action.Action)
-	}
-
-	switch err {
-	case nil:
-		// woo
-	case client.ErrSnapshotSetNotFound, client.ErrSnapshotSnapsNotFound:
-		return NotFound("%v", err)
-	default:
-		return InternalError("%v", err)
-	}
-
-	chg := newChange(st, action.Action+"-snapshot", action.String(), []*state.TaskSet{ts}, affected)
-	chg.Set("api-data", map[string]interface{}{"snap-names": affected})
-	ensureStateSoon(st)
-
-	return AsyncResponse(nil, &Meta{Change: chg.ID()})
 }
